# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Platform V2 API",
...
    "version": "0.5.0"
  },
  "paths": {
...
    "/v1/scans": {
      "post": {
        "tags": [
          "scans"
        ],
        "summary": "Create scan and enqueue collection",
        "description": "Starts a new security scan for the given integration. Enqueues a job on `platform:collect.aws`; workers collect AWS inventory to S3, ingest assets, evaluate policies, and compute CIS score.\n\nRequires **tenant_admin** or **super_admin**. Poll **GET /v1/scans/{id}** until status is `completed`, `completed_with_errors`, or `failed` before reading findings.",
        "operationId": "create_scan_v1_scans_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Scan queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "404": {
            "description": "Integration not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      },
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "List scans",
        "description": "Recent scans for the tenant, newest first.",
        "operationId": "list_scans_v1_scans_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scan list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScanResponse"
                  },
                  "title": "Response List Scans V1 Scans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
...
    "/v1/admin/overview": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Platform ops overview",
        "description": "Cross-tenant counts: tenants, failed/active scans, Redis queue depths. **super_admin only.** Login as ops@drantiq.local.",
        "operationId": "admin_overview_v1_admin_overview_get",
        "responses": {
          "200": {
            "description": "Overview metrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOverviewResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not super_admin"
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
    "/v1/admin/tenants": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List all tenants",
        "description": "Cross-tenant list for admin-ui. **super_admin only.**",
        "operationId": "list_tenants_v1_admin_tenants_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TenantResponse"
                  },
                  "type": "array",
                  "title": "Response List Tenants V1 Admin Tenants Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create tenant",
        "description": "Provision a new customer tenant. Slug must be unique lowercase alphanumeric with hyphens.",
        "operationId": "create_tenant_v1_admin_tenants_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          },
          "409": {
            "description": "Slug exists"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
    "/v1/admin/tenants/{tenant_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get tenant",
        "description": "Single tenant by UUID.",
        "operationId": "get_tenant_v1_admin_tenants__tenant_id__get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          },
          "404": {
            "description": "Tenant not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      },
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Update tenant",
        "description": "Rename or suspend/activate tenant.",
        "operationId": "update_tenant_v1_admin_tenants__tenant_id__patch",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          },
          "400": {
            "description": "No fields"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
    "/v1/admin/tenants/{tenant_id}/memberships": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List tenant memberships",
        "description": "Supabase/OIDC users linked to tenant with roles.",
        "operationId": "list_memberships_v1_admin_tenants__tenant_id__memberships_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MembershipResponse"
                  },
                  "title": "Response List Memberships V1 Admin Tenants  Tenant Id  Memberships Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Provision user membership",
        "description": "Link Supabase user (`auth_issuer` + `auth_subject` from JWT) to tenant. Issuer is typically `https://<project>.supabase.co/auth/v1`.",
        "operationId": "create_membership_v1_admin_tenants__tenant_id__memberships_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
    "/v1/admin/scans": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List scans (cross-tenant)",
        "description": "Recent scans across all tenants. Filter `?status=failed` for ops triage.",
        "operationId": "list_admin_scans_v1_admin_scans_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminScanResponse"
                  },
                  "title": "Response List Admin Scans V1 Admin Scans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
    "/v1/admin/tenants/{tenant_id}/integrations": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List tenant integrations (admin)",
        "description": "Same as tenant-scoped integrations list but callable by super_admin for any tenant.",
        "operationId": "list_tenant_integrations_v1_admin_tenants__tenant_id__integrations_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminIntegrationResponse"
                  },
                  "title": "Response List Tenant Integrations V1 Admin Tenants  Tenant Id  Integrations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
    "/v1/admin/tenants/{tenant_id}/scans": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List scans for tenant (admin)",
        "description": "Recent scans for one tenant from admin-ui tenant detail.",
        "operationId": "list_tenant_scans_v1_admin_tenants__tenant_id__scans_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminScanResponse"
                  },
                  "title": "Response List Tenant Scans V1 Admin Tenants  Tenant Id  Scans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Trigger scan for tenant (admin)",
        "description": "Ops-initiated scan from admin-ui tenant detail. Same pipeline as POST /v1/scans.",
        "operationId": "create_tenant_scan_v1_admin_tenants__tenant_id__scans_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminScanCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Scan queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminScanCreateResponse"
                }
              }
            }
          },
          "404": {
            "description": "Tenant or integration not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScanCreate": {
        "properties": {
          "integration_id": {
            "type": "string",
            "format": "uuid",
            "title": "Integration Id",
            "description": "UUID from GET /v1/integrations"
          }
        },
        "type": "object",
        "required": [
          "integration_id"
        ],
        "title": "ScanCreate"
      },
...
      "ScanResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "integration_id": {
            "type": "string",
            "title": "Integration Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "integration_id",
          "status",
          "trace_id",
          "created_at",
          "updated_at"
        ],
        "title": "ScanResponse"
      },
...
      "TenantResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "TenantResponse"
      },
...
    },
    "securitySchemes": {
      "SupabaseJWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase `access_token` from POST .../auth/v1/token?grant_type=password. Paste token only in Authorize dialog."
      },
      "DevTenantId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tenant-ID",
        "description": "Tenant UUID. Only when DEV_TENANT_HEADER_AUTH=true."
      },
      "DevRole": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Role",
        "description": "tenant_admin | viewer | super_admin. Pair with DevTenantId."
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8090",
      "description": "Local dev"
    },
    {
      "url": "https://api.example.com",
      "description": "Staging / production"
    }
  ]
}